requirements.txt
flask
crewai
crewai-tools
python-dotenv
gunicorn
crewai[google-genai]
